fix: update CAN bootloader CRC and project structure
- Change CRC handling from 32-bit to 16-bit values.
- Update UDS checksum parsing to accept a 16-bit CRC.
- Correct flash sector calculation and erase sequencing.
- Adjust the application flash memory range.
- Rename the `app` directory to `application`.
- Update the CAN bootloader binary.

+++ can_bootloader/Flash_app/inc/fls_app.h
@@ -69,7 +69,7 @@ extern uint8 Flash_IsReadAppInfoFromFlashValid(void);
 extern uint8 Flash_IsAppInFlashValid(void);
 
 /*save received check sum crc*/
-extern void Flash_SavedReceivedCheckSumCrc(uint32 i_receivedCrc);
+extern void Flash_SavedReceivedCheckSumCrc(uint16 i_receivedCrc);
 
 /*erase flash driver in RAM*/
 extern void Flash_EraseFlashDriverInRAM(void);
